clippy optimisations

+++ crates/cli/utils.rs
@@ -32,11 +32,11 @@ mod tests {
     let message = "Hello".repeat(100000);
     let bytes = message.as_bytes();
     let len = bytes.len();
-    assert_eq!(len, 500000 as usize);
+    assert_eq!(len, 500000_usize);
     let to_u8_array = usize_to_u8_array(len as u32);
     let expected: [u8; 4] = [32, 161, 7, 0];
     assert_eq!(to_u8_array, expected);
-    assert_eq!(500000 as usize, u8_array_to_usize(to_u8_array));
+    assert_eq!(500000_usize, u8_array_to_usize(to_u8_array));
     assert_eq!(usize_to_u8_array(500000), expected.to_owned());
   }
 }
